利用mybatis修改了原代码逻辑

+++ backend/src/main/java/com/pbdcompany/dto/request/PaymentRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class PaymentRequest {
-    private Long orderId;
+    private int orderId;
     private String paymentMethod;
     private String verificationCode;
